Refactored to share state and do more logic on app startup (using @vue/reactivity)
